Update ransomware message sig
This sig does not appear to work in original form. It has been modified to ignore completely dropped files which is no longer a possible detection point and merely check for a possible match in a file that is written as a likely ransomware match breaking (so it should get first match only for performance).

Extended also to .txt file to cover most though .htm and html files will need looked at also for possible ransom notes.