Create signature for \x00 wiper malware
Add signature for when potential wiper overwrites multiple files with \x00. This can be seen in CaddyWiper which was detected being used in Ukraine 14/3/2022. 

There are other ways to overwrite files and random bytes would be another but I do not have a detection method for that technique yet although I am thinking along the lines of overwriting multiple files it did not create but this would be a lower severity signature due to it being also indicative of installers.