add loader_alien.py
This PR/commit is an open source contribution from the Proofpoint Threat Detection & Research team as part of the Virus Bulletin 2024 conference paper and presentation (https://www.virusbulletin.com/conference/vb2024/abstracts/life-and-death-building-detection-forensics-and-intelligence-scale/).

It's nothing much, but believe me, it was not easy to get this through the approval process.

AlienLoader
AlienLoader is a Java dropper used by TA577, which was seen dropping Pikabot. Proofpoint researchers dubbed AlienLoader due to the popular threat label name from VirusTotal. Looking at the decompiled binary and its behavior in a sandbox run identifies numerous detection opportunities. The sample reaches out to the payload URL and tries to create and write the downloaded payload to a file under “C:\Users\Public\filename.exe” before it sleeps for 15,000 milliseconds. After that, it starts “filename.exe” in a new process. Chaining this behavior together in a state machine results in a dynamic detection signature for this loader.