Create ThreatFox Intel Sig
This is an experimental signature for ThreatFox. It still needs the following assistance/work but if the api is entered from your threatfox account (free to create) this will work on DNS requests. 

So what is needed:
- I need to add in lookups for other network types (IP connections and other types if necessary)

The main thing is ideally is rather than the API being entered in the signature it is entered in processing.conf. I have created the following in my processing.conf under community:

[threatfox]
enabled = no
# Get your API key from https://threatfox.abuse.ch/ after creating a free account.
apikey = ""

My question is how do a call that checking that it is enabled and then more specifically how do I then pull the apikey from this into the signature.

So I think something like:
processing_conf = Config("processing")

if processing_conf.threatfox.enabled:
    api = GET_APIKEY (NEED HELP HERE)
    if api:
        rest of signature