Create signature for DLL sideloading
Initial signature for one behaviour that appears in some sideloading cases. For this to trigger either zip_compound needs to be used triggering the correct file that then sideloads the DLL or within the chain naturally (i.e. a loader/malicious script/doc etc. pulls down the elements needed itself and sideloads it. 

Some other sigs will need investigated for anomalies on this which I am looking into but this should detect some cases.