Update network_dns_suspicious.py
Necessary in order to actually hit on the actual domains, the point need to be escaped or it can be any character. Also the domain need to be finishing there so it is not as a subdomain or else it is not a shortener. This should be a domain not an URI so no issue for this.