NEEDS TESTING: Syscall Signature (fresh ntdll load)
So this needs testing please as I don't have a current CAPE until i am back in office/have new home lab PC in a few weeks. 

The logic of this is that the Windows loader should take care of NTDLL. Loading a fresh NTDLL into itself from disk likely is then to begin parsing it for syscall instruction for doing that or to bypass hooking.

Here are two samples that do this for testing each element: 

1) AgentTesla https://capesandbox.com/analysis/46120/ 

This uses NTCreateFile on ntdll to load it and then NtReadFile to read it into memory.

2) LUMMA https://capesandbox.com/analysis/46359/

This uses NTCreateFile on ntdll to load it then maps in into its memory with NTCreateSection before changing the memory protections on it (likely to then patch it and remove hooks).